+++ packages/creem-sdk/RELEASES.md
@@ -79,3 +79,13 @@ Based on:
 - [typescript v1.5.3] packages/creem-sdk
 ### Releases
 - [NPM v1.5.3] https://www.npmjs.com/package/creem/v/1.5.3 - packages/creem-sdk
+
+## 2026-07-15 09:30:53
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.790.1 (2.918.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v1.5.4] packages/creem-sdk
+### Releases
+- [NPM v1.5.4] https://www.npmjs.com/package/creem/v/1.5.4 - packages/creem-sdk